Move CERC API forecast fixtures into own namespace
We are going to build test fixtures for our own `Forecast`
entities as distinct from the JSON representations which we
receive from CERC's `/getforecast/all?` endpoint.
