+++ spec/feature_steps/forecast_steps.rb
@@ -1,14 +1,14 @@
 module ForecastSteps
   def given_a_forecast_for_today
-    forecasts << Fixtures.build_forecast(day: :today, air_pollution_status: :high, uv_level: 1, pollen_level: 4, min_temperature: 0.5, max_temperature: 2.4)
+    forecasts << Fixtures::API.build_forecast(day: :today, air_pollution_status: :high)
   end
 
   def and_a_forecast_for_tomorrow
-    forecasts << Fixtures.build_forecast(day: :tomorrow, air_pollution_status: :moderate, uv_level: 2, pollen_level: 5, min_temperature: 2.7, max_temperature: 4.0)
+    forecasts << Fixtures::API.build_forecast(day: :tomorrow, air_pollution_status: :moderate)
   end
 
   def and_a_forecast_for_the_day_after_tomorrow
-    forecasts << Fixtures.build_forecast(day: :day_after_tomorrow, air_pollution_status: :very_high, uv_level: 3, pollen_level: 6, min_temperature: 5.2, max_temperature: 5.9)
+    forecasts << Fixtures::API.build_forecast(day: :day_after_tomorrow, air_pollution_status: :very_high)
   end
 
   def forecasts
@@ -16,7 +16,7 @@ module ForecastSteps
   end
 
   def and_the_response_from_cercs_api_is_stubbed_accordingly
-    forecast_response = Fixtures.forecast_wrapper_around(@forecasts)
+    forecast_response = Fixtures::API.forecast_wrapper_around(@forecasts)
     HttpStubs.stub_forecasts_with(forecast_response)
   end
